[🚚 move] 패키지 구조변경
