--- com.hub-service/src/main/java/com/hub_service/application/dto/HubCreateRequestDto.java
@@ -1,4 +1,0 @@
-package com.hub_service.application.dto;
-
-public class HubCreateRequestDto {
-}
