--- com.hub-service/src/main/java/com/hub_service/config/RedisConfig.java
@@ -1,4 +1,0 @@
-package com.hub_service.config;
-
-public class RedisConfig {
-}
