--- com.hub-service/src/main/java/com/hub_service/domain/entity/Hub.java
@@ -1,4 +1,0 @@
-package com.hub_service.domain.entity;
-
-public class Hub {
-}
