+++ com.hub-service/src/main/java/com/hub_service/domain/model/Hub.java
@@ -1,0 +1,4 @@
+package com.hub_service.domain.model;
+
+public class Hub {
+}
