--- com.hub-service/src/main/java/com/hub_service/exception/GlobalExceptionHandler.java
@@ -1,4 +1,0 @@
-package com.hub_service.exception;
-
-public class GlobalExceptionHandler {
-}
