+++ com.hub-service/src/main/java/com/hub_service/infrastructure/config/SecurityConfig.java
@@ -1,0 +1,4 @@
+package com.hub_service.infrastructure.config;
+
+public class SecurityConfig {
+}
