--- com.hub-service/src/main/java/com/hub_service/presentation/HubController.java
@@ -1,4 +1,0 @@
-package com.hub_service.presentation;
-
-public class HubController {
-}
