+++ com.hub-service/src/main/java/com/hub_service/presentation/advice/GlobalExceptionHandler.java
@@ -1,0 +1,4 @@
+package com.hub_service.presentation.advice;
+
+public class GlobalExceptionHandler {
+}
