+++ com.hub-service/src/main/java/com/hub_service/presentation/controller/HubController.java
@@ -1,0 +1,4 @@
+package com.hub_service.presentation.controller;
+
+public class HubController {
+}
