+++ com.hub-service/src/main/java/com/hub_service/presentation/dto/request/HubCreateRequestDto.java
@@ -1,0 +1,4 @@
+package com.hub_service.presentation.dto.request;
+
+public class HubCreateRequestDto {
+}
