+++ com.hub-service/src/main/java/com/hub_service/presentation/dto/response/HubResponseDto.java
@@ -1,0 +1,4 @@
+package com.hub_service.presentation.dto.response;
+
+public class HubResponseDto {
+}
